+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog    
+## Release v0.1.2-alpha `04.03.2019`
+* Update of README.md, [About Slingshot](https://github.com/voodooEntity/slingshotdb/blob/master/docs/ABOUT_SLINGSHOT.md) 
+* Adding 'context' filter filter param to multiple methods
+    * `getChildEntities` (applies on the child entity not the relation) 
+    * `getParentEntities` (applies on the parental entity not the relation)
+    * `getRelationsTo` (applies on the relation)
+    * `getRelationsFrom` (applies on the relation)
+    * `getEntitiesByValue` 
+* Fixing a possible concurrency bug in `storage.GetParentRelationsByTargetTypeAndTargetId`
+* Little comment cleaning in storage
+
 ## Release v0.1.2-alpha `21.12.2019`    
 * Update of README.md, [About Slingshot](https://github.com/voodooEntity/slingshotdb/blob/master/docs/ABOUT_SLINGSHOT.md) and [HTTP API](https://github.com/voodooEntity/slingshotdb/blob/master/docs/HTTP_API_V1.md)
 * Adding optional param `mode` string to `getEntitiesByValue` and `getEntitiesByTypeAndValue` to decide wich compare mode should be used. Available now are